format: Improve handling of commit authorship and mail "From:" headers
Default to the behavior of git format-patch --from, which uses your
committer identity as the "From:" header for the patch mail, and
preserves commit authorship in an in-body "From:" header when formatting
commits you didn't write.  Add a --no-from option to use the commit
author as the "From:" address for each patch mail.