contrib: Add bash completions
I'm definitely not a bash completions wizard, but it was annoying me
enough to scratch my own itch. These completions are fairly completely
(req isn't implemented, but git doesn't have completions for
pull-request either).

This is based on git bash completions included in git (and in fact rely
on them) making them a derivative work, and therefore GPLv2.

Fixes #20