Support diffs between patch series
Inspired by tbdiff, the diff algorithm computes the interdiff for each
possible corresponding pair of commits (as well as the possibility of an
added or deleted commit), and then treats that as an instance of the
Assignment Problem to find a correspondence with the minimal total
interdiffs.

Add a new `git series diff` command to show the diff between the working
and staged versions of the patch series.  Also use the new diff format
for `git series log -p`.